requirements.txt
# Pour l'analyse de données
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2

# Pour la génération de PDF
reportlab==3.6.13
weasyprint==60.1

# Pour le traitement et l'affichage des graphiques
plotly==5.18.0

# Pour les tâches périodiques
celery==5.3.4
django-celery-beat==2.5.0